move counters closer to where they are used
